Remove hard coded database names where external Postgres connections will be used. Move env var handling to helpers.
